+++ pyriemann/utils/ajd.py
@@ -31,7 +31,7 @@ def rjd(X, threshold=1e-8, n_iter_max=1000):
 
     See Also
     --------
-    wedge
+    ajd_pham
 
     References
     ----------
@@ -85,5 +85,89 @@ def rjd(X, threshold=1e-8, n_iter_max=1000):
                     V[:, p] = c * V[:, p] + s * V[:, q]
                     V[:, q] = c * V[:, q] - s * tmp
 
+    D = np.reshape(A, (m, nm/m, m)).transpose(1, 0, 2)
+    return V, D
+
+
+def ajd_pham(X, threshold=1e-6, n_iter_max=15):
+    """Approximate joint diagonalization based on pham's algorithm.
+
+    This is a direct implementation of the PHAM's AJD algorithm.
+
+    Parameters
+    ----------
+    X : ndarray, shape (n_trials, n_channels, n_channels)
+        A set of covariance matrices to diagonalize
+    threshold : float (default 1e-8)
+        The number of standard deviation to reject artifacts.
+    n_iter_max : int (default 1000)
+        The maximum number of iteration to reach convergence.
+
+    Returns
+    -------
+    V : ndarray, shape (n_channels, n_channels)
+        the diagonalizer
+    D : ndarray, shape (n_trials, n_channels, n_channels)
+        the set of quasi diagonal matrices
+
+    Notes
+    -----
+    .. versionadded:: 0.2.4
+
+    See Also
+    --------
+    rjd
+
+    """
+    nmat = X.shape[0]
+
+    # reshape input matrix
+    A = np.concatenate(X, 0).T
+
+    # init variables
+    m, nm = A.shape
+    V = np.eye(m)
+    epsi = m * (m - 1)*threshold
+
+    for it in range(n_iter_max):
+        decr = 0
+        for i in range(1, m):
+            for j in range(i):
+                Ii = np.arange(i, nm, m)
+                Ij = np.arange(j, nm, m)
+
+                c1 = A[i, Ii]
+                c2 = A[j, Ij]
+
+                g12 = np.mean(A[i, Ij] / c1)
+                g21 = np.mean(A[i, Ij] / c2)
+
+                omega21 = np.mean(c1 / c2)
+                omega12 = np.mean(c2 / c1)
+                omega = np.sqrt(omega12*omega21)
+
+                tmp = np.sqrt(omega21/omega12)
+                tmp1 = (tmp*g12 + g21)/(omega + 1)
+                tmp2 = (tmp*g12 - g21)/np.max(omega - 1, 1e-9)
+
+                h12 = tmp1 + tmp2
+                h21 = np.conj((tmp1 - tmp2)/tmp)
+
+                decr = decr + nmat*(g12 * np.conj(h12) + g21 * h21) / 2.0
+
+                tmp = 1 + 1.j * 0.5 * np.imag(h12 * h21)
+                tmp = np.real(tmp + np.sqrt(tmp ** 2 - h12 * h21))
+                T = np.array([[1, -h12/tmp], [-h21/tmp, 1]])
+
+                A[[i, j], :] = np.dot(T, A[[i, j], :])
+                tmp = np.c_[A[:, Ii], A[:, Ij]]
+                tmp = np.dot(np.reshape(tmp, (m * nmat, 2), order='F'), T.T)
+
+                tmp = np.reshape(tmp, (m, nmat * 2), order='F')
+                A[:, Ii] = tmp[:, :nmat]
+                A[:, Ij] = tmp[:, nmat:]
+                V[[i, j], :] = np.dot(T, V[[i, j], :])
+        if decr < epsi:
+            break
     D = np.reshape(A, (m, nm/m, m)).transpose(1, 0, 2)
     return V, D