+++ doc/whatsnew.rst
@@ -22,9 +22,9 @@ v0.2.4 (Unreleased)
 
 - Added CSP for covariance matrices.
 
-- Added Approximate Joint diagonalization algorithm (JADE)
+- Added Approximate Joint diagonalization algorithm (JADE, PHAM)
 
-- Added ALE mean 
+- Added ALE mean
 
 v0.2.3 (November 2015)
 ----------------------